backend/requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
django-cors-headers>=4.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
whitenoise>=6.7
gunicorn>=22.0
dj-database-url>=2.2
psycopg2-binary>=2.9
django-storages[boto3]>=1.14
boto3>=1.34
